changes merged

+++ lib/data/models/response/getRatingsAndReview.dart
@@ -11,19 +11,20 @@ class GetRatingsAndReview {
     this.success,
   });
 
-  factory GetRatingsAndReview.fromJson(Map<String, dynamic> json) => GetRatingsAndReview(
-    message: json["message"],
-    data: json["data"] == null ? null : Data.fromJson(json["data"]),
-    statusCode: json["statusCode"],
-    success: json["success"],
-  );
+  factory GetRatingsAndReview.fromJson(Map<String, dynamic> json) =>
+      GetRatingsAndReview(
+        message: json["message"],
+        data: json["data"] == null ? null : Data.fromJson(json["data"]),
+        statusCode: json["statusCode"],
+        success: json["success"],
+      );
 
   Map<String, dynamic> toJson() => {
-    "message": message,
-    "data": data?.toJson(),
-    "statusCode": statusCode,
-    "success": success,
-  };
+        "message": message,
+        "data": data?.toJson(),
+        "statusCode": statusCode,
+        "success": success,
+      };
 }
 
 class Data {
@@ -34,17 +35,19 @@ class Data {
   });
 
   factory Data.fromJson(Map<String, dynamic> json) => Data(
-    reviewRating: json["reviewRating"] == null ? null : ReviewRating.fromJson(json["reviewRating"]),
-  );
+        reviewRating: json["reviewRating"] == null
+            ? null
+            : ReviewRating.fromJson(json["reviewRating"]),
+      );
 
   Map<String, dynamic> toJson() => {
-    "reviewRating": reviewRating?.toJson(),
-  };
+        "reviewRating": reviewRating?.toJson(),
+      };
 }
 
 class ReviewRating {
   List<Review>? reviews;
-  int? ratingAvg;
+  double? ratingAvg;
 
   ReviewRating({
     this.reviews,
@@ -52,14 +55,19 @@ class ReviewRating {
   });
 
   factory ReviewRating.fromJson(Map<String, dynamic> json) => ReviewRating(
-    reviews: json["reviews"] == null ? [] : List<Review>.from(json["reviews"]!.map((x) => Review.fromJson(x))),
-    ratingAvg: json["ratingAvg"],
-  );
+        reviews: json["reviews"] == null
+            ? []
+            : List<Review>.from(
+                json["reviews"]!.map((x) => Review.fromJson(x))),
+        ratingAvg: json["ratingAvg"],
+      );
 
   Map<String, dynamic> toJson() => {
-    "reviews": reviews == null ? [] : List<dynamic>.from(reviews!.map((x) => x.toJson())),
-    "ratingAvg": ratingAvg,
-  };
+        "reviews": reviews == null
+            ? []
+            : List<dynamic>.from(reviews!.map((x) => x.toJson())),
+        "ratingAvg": ratingAvg,
+      };
 }
 
 class Review {
@@ -86,26 +94,26 @@ class Review {
   });
 
   factory Review.fromJson(Map<String, dynamic> json) => Review(
-    createdAt: json["createdAt"],
-    contentId: json["contentId"],
-    rating: json["rating"],
-    comment: json["comment"],
-    title: json["title"],
-    reviewId: json["reviewId"],
-    userId: json["userId"],
-    userProfile: json["userProfile"],
-    username: json["username"],
-  );
+        createdAt: json["createdAt"],
+        contentId: json["contentId"],
+        rating: json["rating"],
+        comment: json["comment"],
+        title: json["title"],
+        reviewId: json["reviewId"],
+        userId: json["userId"],
+        userProfile: json["userProfile"],
+        username: json["username"],
+      );
 
   Map<String, dynamic> toJson() => {
-    "createdAt": createdAt,
-    "contentId": contentId,
-    "rating": rating,
-    "comment": comment,
-    "title": title,
-    "reviewId": reviewId,
-    "userId": userId,
-    "userProfile": userProfile,
-    "username": username,
-  };
+        "createdAt": createdAt,
+        "contentId": contentId,
+        "rating": rating,
+        "comment": comment,
+        "title": title,
+        "reviewId": reviewId,
+        "userId": userId,
+        "userProfile": userProfile,
+        "username": username,
+      };
 }
